# src/main/resources/swagger/rest-transfer-service-swagger.yml
swagger: "2.0"
info:
  description: "This is project assignment"
  version: "1.0.0"
  title: "TransferService API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "suhabb22@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8090"
basePath: "/rest-transfer-service"
tags:
- name: "transfer"
  description: "Transfer Amount Between Accounts"
- name: "account"
  description: "Account Info"
schemes:
- "http"
paths:
  /transfer:
    post:
      tags:
      - "transfer"
      summary: "TransferAmount Between Accounts"
      description: ""
      operationId: "transfer"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Transfer Object"
        required: true
        schema:
          $ref: "#/definitions/Transfer"
      responses:
        "400":
          description: "Invalid input"
  /account/{accountNumber}:
    get:
      tags:
      - "account"
      summary: "Get Account by Account Number"
      operationId: "getAccountByAccountNumber"
      produces:
      - "application/json"
      parameters:
      - name: "accountNumber"
        in: "path"
        description: "Account number to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Account"
        "400":
          description: "Invalid ID supplied"
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: "Account not found"
definitions:
  Tag:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Tag"
  Transfer:
    type: "object"
    required:
    - "sourceAccount"
    - "destinationAccount"
    - "amount"
    properties:
      sourceAccount:
        type: "integer"
        format: "int64"
      destinationAccount:
        type: "integer"
        format: "int64"
      amount:
        type: "number"
        format: "double"
  Account:
    type: "object"
    required:
      - "accountNumber"
      - "balance"
    properties:
      accountNumber:
        type: "integer"
        format: "int64"
      amount:
        type: "number"
        format: "double"
    Error:
      type: "object"
      required:
        - "status"
        - "title"
        - "description"
      properties:
        status:
          type: "string"
        title:
          type: "string"
        description:
          type: "string"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"